requirements.txt
asgiref==3.7.2
Django
psycopg2-binary
pillow
ipython
djangorestframework